Conflicts:
	modules/Makefile

+++ runtime/src/qio/qio.c
@@ -3587,7 +3587,7 @@ void _qio_channel_write_bits_cached_realign(qio_channel_t* restrict ch, uint64_t
   tmp_bits = ch->bit_buffer;
 
   // ch->bit_buffer_bits should never exceed the sizeof the bitbuffer
-  assert(0 <= tmp_live && tmp_live <= 8*sizeof(qio_bitbuffer_t));
+  assert(0 <= tmp_live && tmp_live <= (int) (8*sizeof(qio_bitbuffer_t)));
 
   // We've got > 64 bits to write.
   part_one = 8*sizeof(qio_bitbuffer_t) - tmp_live;
@@ -3665,7 +3665,7 @@ qioerr _qio_channel_write_bits_slow(qio_channel_t* restrict ch, uint64_t v, int8
   if( tmp_live == 0 ) tmp_bits = 0;
 
   // ch->bit_buffer_bits should never exceed the sizeof the bitbuffer
-  assert(0 <= tmp_live && tmp_live <= 8*sizeof(qio_bitbuffer_t));
+  assert(0 <= tmp_live && tmp_live <= (int) (8*sizeof(qio_bitbuffer_t)));
 
   //printf("In write bits slow\n");
